Rename VolumeScene
