+++ VolumeRendering/Main.cpp
@@ -1,5 +1,5 @@
 #include <App.h>
-#include "VolumeScene.h"
+#include "RayMarchingScene.h"
 int main(int argc, char* argv[]) {
 	std::shared_ptr<RayMarchingScene> scene = std::make_shared<RayMarchingScene>();
 	App::onInitialization(scene, argc, argv);
