+++ VolumeRendering/VolumeRendering.vcxproj
@@ -163,7 +163,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Main.cpp" />
-    <ClCompile Include="VolumeScene.cpp" />
+    <ClCompile Include="RayMarchingScene.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Image Include="body-at_4096.jpg" />
@@ -185,7 +185,7 @@
     <None Include="volumeCubeVS.vert" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="VolumeScene.h" />
+    <ClInclude Include="RayMarchingScene.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
