+++ VolumeRendering/VolumeRendering.vcxproj.filters
@@ -24,7 +24,7 @@
     <ClCompile Include="Main.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="VolumeScene.cpp">
+    <ClCompile Include="RayMarchingScene.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -78,7 +78,7 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="VolumeScene.h">
+    <ClInclude Include="RayMarchingScene.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
